ProjectStatusesFunctionality

+++ Server/Models/ProductionStatus.cs
@@ -4,7 +4,7 @@ namespace Server.Models
 {
     public class ProductionStatus
     {
-        public byte ProductionStatusID { get; set; }
+        public int ProductionStatusID { get; set; }
         public string ProductionStatusName { get; set; }
 
         public ProductionStatus() { }
